Moving the translation resources into a separate directory.
