+++ grow/translations/catalogs_test.py
@@ -3,9 +3,9 @@
 import os
 import unittest
 from grow.pods import pods
-from grow.pods import catalogs
 from grow.pods import storage
 from grow.testing import testing
+from grow.translations import catalogs
 
 
 class CatalogTest(unittest.TestCase):
