+++ grow/translations/importers_test.py
@@ -1,11 +1,13 @@
+"""Tests for importers."""
+
 import os
+import unittest
 from grow.pods import pods
 from grow.pods import storage
 from grow.testing import testing
-import unittest
 
 
-class CatalogsTest(unittest.TestCase):
+class ImportersTest(unittest.TestCase):
 
     def setUp(self):
         dir_path = testing.create_test_pod_dir()
